git-svn-id: svn://123.57.173.36/jhj-parent2.0/Branches/jhj-parent2.1@420 57d5540f-7636-4a73-8af3-7dc9f519f443

+++ jhj-dao/src/main/java/com/jhj/po/model/survey/SurveyContent.java
@@ -27,7 +27,7 @@ public class SurveyContent {
 
     private Short contentChildType;
 
-    private Integer defaultTime;
+    private Long defaultTime;
 
     public Long getContentId() {
         return contentId;
@@ -125,11 +125,11 @@ public class SurveyContent {
         this.contentChildType = contentChildType;
     }
 
-    public Integer getDefaultTime() {
+    public Long getDefaultTime() {
         return defaultTime;
     }
 
-    public void setDefaultTime(Integer defaultTime) {
+    public void setDefaultTime(Long defaultTime) {
         this.defaultTime = defaultTime;
     }
 }
